pyproject.toml
[tool.poetry]
name = "ray-memory-dashboard"
version = "1.0.0"
description = "Advanced dashboard for Ray's consciousness and memory system"
authors = ["Ray Consciousness Project"]
readme = "README.md"
packages = [{include = "ui"}, {include = "services"}, {include = "extract"}]

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
streamlit = ">=1.28.0,<2.0.0"
streamlit-ace = ">=0.1.1"
streamlit-extras = ">=0.3.0"
plotly = ">=5.15.0,<6.0.0"
pandas = ">=1.5.0,<3.0.0"
pygments = ">=2.15.0"
sentence-transformers = ">=2.2.0,<3.0.0"
faiss-cpu = ">=1.7.4,<2.0.0"
numpy = ">=1.21.0,<2.0.0"
scikit-learn = ">=1.3.0,<1.7.0"
google-generativeai = "^0.8.5"
chardet = "^5.2.0"
fastapi = "^0.116.1"
uvicorn = "^0.35.0"
requests = "^2.32.4"
psutil = "^7.0.0"

[tool.poetry.group.dev.dependencies]
pytest = ">=7.0.0,<8.0.0"
black = ">=23.0.0,<24.0.0"
flake8 = ">=6.0.0,<7.0.0"
mypy = ">=1.0.0,<2.0.0"

[tool.poetry.scripts]
dashboard = "run_dashboard:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true